+++ lib/data/models/styles_models/surah_name_style.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart' show Color;
 
 class SurahNameStyle {
-  final Color surahNameColor;
+  final Color? surahNameColor;
   final double? surahNameWidth;
   final double? surahNameHeight;
 
   SurahNameStyle({
-    required this.surahNameColor,
+    this.surahNameColor,
     this.surahNameWidth,
     this.surahNameHeight,
   });